switch out itao for a standard way (function provided by @pharab)

+++ source/waternet/src/printfuncs.h
@@ -5,7 +5,7 @@
 
 void printDebugCpuLoad();
 void printDebug(uint8_t ax, uint8_t ay, const __FlashStringHelper* amsg);
-void printNumber(uint8_t ax, uint8_t ay, uint16_t aNumber, uint8_t maxDigits);
+void printNumber(uint8_t ax, uint8_t ay, uint16_t aNumber, size_t maxDigits);
 void printMessage(uint8_t ax, uint8_t ay, const __FlashStringHelper* amsg);
 void printCongratsScreen(uint8_t ax, uint8_t ay, const __FlashStringHelper* amsg);
